pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "asset_worksheet"
version = "0.0.1"
authors = [
  { name="Jamie Duncan", email="jamieduncan@google.com" },
]
description = "Generates an automated inventory worksheets from GCP Asset Inventory APIs. Current Supported: FedRAMP Attachment 13"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    'License :: OSI Approved :: Apache Software License',
    "Operating System :: OS Independent"
]

dependencies = [
  "google-cloud-asset",
  "google-cloud-storage"
]

[project.urls]
"Homepage" = "https://github.com/jduncan-rva/GCPAttachment13"
"Bug Tracker" = "https://github.com/jduncan-rva/GCPAttachment13/issues"

[project.scripts]
assetWorksheet = "assetWorksheet:createWorksheet"